Don't hide exception when download fails

+++ extension-manager-integration-test-java/src/main/java/com/exasol/extensionmanager/itest/PreviousVersionManager.java
@@ -59,7 +59,9 @@ public class PreviousVersionManager {
             downloadToFile(url, tempFile);
             uploadToBucketFs(fileName, tempFile);
         } finally {
-            deleteFile(tempFile);
+            if (Files.exists(tempFile)) {
+                deleteFile(tempFile);
+            }
         }
     }
